TrabalhoServerEntities/src/main/java/br/edu/ifes/sr/poo2/model/Servico.java
package br.edu.ifes.sr.poo2.model;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.OneToOne;

@Entity
public class Servico extends Model{
	
	/**
	 * 
	 */
	private static final long serialVersionUID = 1L;

	@Column
	private String url;
	
	@OneToOne
	private Categoria categoria;
	
	public Categoria getCategoria() {
		return categoria;
	}

	public void setCategoria(Categoria categoria) {
		this.categoria = categoria;
	}

	public static long getSerialversionuid() {
		return serialVersionUID;
	}

	public String getUrl() {
		return url;
	}

	public void setUrl(String url) {
		this.url = url;
	}

	
	
	
}
